requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

pydantic>=2.7.0
pydantic-settings>=2.2.0

google-cloud-bigquery>=3.25.0
google-auth>=2.30.0

python-dotenv>=1.0.1
PyYAML>=6.0.1
